refactor: moves todos into a separate tab
